Prevent calling String.repeat() with a negative number in dev console
If the number of columns is less than the length of the text string, we
would previously call '='.repeat() with a negative number, which
crashes. So, pass 0 if it would be negative.

This is not needed for the generateEnd() function below, because the
number of columns can't be less than 0.

Fixes #203, but doesn't explain why that had a column count of 0.