fix: Don't include non-existent apps in suggest_app_for_fsentry results
On self-hosted installs, several of these apps are not (currently)
available, meaning that `get_app()` would fail and we would get null
values in the suggested_apps array. Since the `/open_file` API endpoint
selects the first result, it could therefore think that there were no
available apps, when actually there were.

For example, source code files can be opened either in `code` or
`editor`. `code` is not available on self-hosted, so the first entry
would be null. This change removes that, so `editor` is rightly
selected.

Also, we have this `file_extension` variable, so let's use it. :^)