feat(phoenix): Add an `open` utility to Phoenix
`open FILE|URL` opens the given file in the user's preferred app.

On Node, we rely on the system's built-in utility rather than attempt to
write our own code for figuring out what the preferred application is.

On Puter, `puter.ui.launchApp()` can't yet grant access to files, so
using `open` on Puter launches the app without the file open.

The other issue is that we need to make the app-launching code from the
CommandProviders available to coreutils and elsewhere, so that they can
interact with stdio without duplicating all the logic.