fix: access to undefined symbol
Lines accessing `model` instead of `options.model` were causing an
exception to be thrown. This was not detected earlier because
`max_tokens` is usually already provided and we don't reach the
conditional expression that does this. However, this code was tested on
a different branch that erroneously skips this logic which brought this
bug to surface.