refactor: re-implement puter_path coerce and cleanup
This commit has cross-concern because I forgot to commit the cleanup
portion previously, and separating the changes did not seem worthwhile.
This could be done retroactively (i.e. during a rebase) but that's
probably not worthwhile either because this is likely to be squash
merged.

This commit implements coercion of puter_path using the interface
provided by the new module (content parts with type=data).