dev: add "how to get X from Y" rules (gateway.js)o
This commit adds a registry for functions that implement whatever
behavior is necessary to produce a specified output value with the given
input values. This simplifies a few different things:
- trims off unnecessary constructs; now everything can be one of these
  definitions
- makes it possible to have different entrypoints without adding more
  code, so Puter can continue to pass its own `completionWriter`

Potential caveat is that the way the an interaction with an LLM is
invoked using this interface is unconventional compared to existing SDKs
, but this doesn't need to be the interface a consumer of this module
uses; we can employ the facade pattern to make it easier to use.