dev: add vibe-coded doc generator; might be replaced
Claude didn't do a great job, but it didn't do a terrible job. I'm
committing this code to give it a little bit of time in the limelight to
see if it's something I can make work. For now, it's promising, but if
it turns out to be more grief than it's worth it will be removed
promptly and without hesitation.