clean: remove ConfigurableCountingService
The ConfigurableCountingService class was only used prior to the current
implementation of metering. The service was registered as 'counting'. To
verify that the class was no longer in use I checked for the string
"ConfigurableCountingService" to verify that the only import was in
CoreModule.js. Then I searched for `'counting`, `"counting"` and
`counting`, finding no matches where the service was being obtained from
the registry. This is strong evidence that this code is no longer used
anywhere. Additionally, this code should no longer be used anywhere.