fix: pass exporter to NodeSDK constructor
Previously a new ConsoleSpanExporter was passed to the constructor of
NodeSDK. This was possibly unintentional but it did not cause any
issues. After the recent update to opentelemetry, or perhaps the changes
introduced in 3c579dcc0d971327a49d9cf290decedbdee8329d, this no longer
works as it prevents tools like Jaeger from getting any spans from
Puter's backend.